# api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.0 DO NOT EDIT.
package api

const (
	BearerScopes = "Bearer.Scopes"
)

// Branch defines model for Branch.
type Branch struct {
	CurrentUsers *int32   `json:"currentUsers,omitempty"`
	GroupIds     *[]int32 `json:"groupIds"`
	Id           *int32   `json:"id,omitempty"`
	MaxUsers     *int32   `json:"maxUsers,omitempty"`
	Name         *string  `json:"name"`
}

// BranchCreationDto defines model for BranchCreationDto.
type BranchCreationDto struct {
	MaxUsers *int32  `json:"maxUsers,omitempty"`
	Name     *string `json:"name"`
}

// BranchDeleteDto defines model for BranchDeleteDto.
type BranchDeleteDto struct {
	BranchIds *[]int32 `json:"branchIds"`
}

// UpdateBranchDto defines model for UpdateBranchDto.
type UpdateBranchDto struct {
	MaxUsers *int32  `json:"maxUsers"`
	Name     *string `json:"name"`
}

// CheckBranchLimitParams defines parameters for CheckBranchLimit.
type CheckBranchLimitParams struct {
	UsersAmount *int32 `form:"usersAmount,omitempty" json:"usersAmount,omitempty"`
	BranchId    *int32 `form:"branchId,omitempty" json:"branchId,omitempty"`
}

// DeleteBranchJSONRequestBody defines body for DeleteBranch for application/json ContentType.
type DeleteBranchJSONRequestBody = BranchDeleteDto

// CreateBranchJSONRequestBody defines body for CreateBranch for application/json ContentType.
type CreateBranchJSONRequestBody = BranchCreationDto

// UpdateBranchJSONRequestBody defines body for UpdateBranch for application/json ContentType.
type UpdateBranchJSONRequestBody = UpdateBranchDto