# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

type Branch struct {
	ID           int32   `json:"id"`
	Name         string  `json:"name"`
	MaxUsers     int32   `json:"maxUsers"`
	CurrentUsers int32   `json:"currentUsers"`
	GroupIds     []int32 `json:"groupIds"`
}